Added new module - NetworkModule
