+++ src/main/java/InfrastructureManager/Modules/NetworkStructure/Exception/NetworkModuleException.java
@@ -1,0 +1,13 @@
+package InfrastructureManager.Modules.NetworkStructure.Exception;
+
+import InfrastructureManager.ModuleManagement.Exception.Execution.ModuleExecutionException;
+
+public class NetworkModuleException extends ModuleExecutionException {
+    public NetworkModuleException(String errorMessage){
+        super(errorMessage);
+    }
+
+    public NetworkModuleException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
