+++ src/main/java/InfrastructureManager/Modules/NetworkStructure/LocationConnectionRelation.java
@@ -1,0 +1,59 @@
+package InfrastructureManager.Modules.NetworkStructure;
+/**
+ * This class contains relationship between location and connection information.
+ *
+ * @author Shankar Lokeshwara
+ */
+public class LocationConnectionRelation {
+	
+	private Connection connection;
+	private Location location;
+	
+	/**
+	* Parameterized constructor for LocationConnectionRelation Class
+	* @param connection 	
+	* @param location 			
+	*/
+	public LocationConnectionRelation(Connection connection,Location location) {
+		this.connection = connection;
+		this.location = location;
+	}
+	
+	/**
+	 * Getter function
+	 * @return location
+	 */
+	public Location getLocationRelation () {
+		return this.location;	
+	}
+	
+	/**
+	 * Getter function
+	 * @return connection
+	 */
+	public Connection getConnectionRelation () {
+		return this.connection;	
+	}
+	
+
+	//ToDo move it to other file with other function implementation
+	//source - https://www.geodatasource.com/developers/java and https://www.w3resource.com/java-exercises/basic/java-basic-exercise-36.php
+	public double calculateDistance(DeviceLocationRelation source,DeviceLocationRelation destination)
+	{
+		double earthRadius = 6371.01; //Kilometers
+		double distance = 0;
+		if ((source.getLocation().getLatitude() == destination.getLocation().getLatitude()) && (source.getLocation().getLongitude() == destination.getLocation().getLongitude())) {
+			distance =  0;
+			}
+
+		else {
+			double theta = source.getLocation().getLongitude() - destination.getLocation().getLongitude();
+			double sourceLatitudeRadians = Math.toRadians(source.getLocation().getLatitude());
+			double destinationLatitudeRadians = Math.toRadians(destination.getLocation().getLatitude());
+	        distance = Math.acos(Math.sin(sourceLatitudeRadians)*Math.sin(destinationLatitudeRadians) + Math.cos(sourceLatitudeRadians)*Math.cos(destinationLatitudeRadians)*Math.cos(Math.toRadians(theta)));
+	        distance = earthRadius * distance;     
+	        }
+		return distance;
+	}
+	
+}
